added conditional to drop tables migration to only drop table if the companies table exists

+++ db/migrate/20140308215112_drop_company_table.rb
@@ -1,6 +1,6 @@
 class DropCompanyTable < ActiveRecord::Migration
   def up
-  	drop_table :companies
+  	drop_table :companies if ActiveRecord::Base.connection.table_exists? 'companies'
   end
 
   def down
